[#79] move ActivityAt to common package
