+++ src/main/java/depromeet/lessonfour/server/activityrecord/app/service/CreateActivityRecordUseCase.java
@@ -9,10 +9,10 @@ import depromeet.lessonfour.server.activityrecord.app.dto.request.CreateActivity
 import depromeet.lessonfour.server.activityrecord.app.support.MealFoodFactory;
 import depromeet.lessonfour.server.activityrecord.domain.entity.ActivityRecord;
 import depromeet.lessonfour.server.activityrecord.domain.repository.ActivityRecordRepository;
-import depromeet.lessonfour.server.activityrecord.domain.vo.ActivityAt;
 import depromeet.lessonfour.server.activityrecord.domain.vo.MealFood;
 import depromeet.lessonfour.server.common.annotation.UseCase;
 import depromeet.lessonfour.server.common.api.code.ErrorCode;
+import depromeet.lessonfour.server.common.domain.vo.ActivityAt;
 import depromeet.lessonfour.server.common.exception.ServerException;
 import lombok.RequiredArgsConstructor;
